Update randomize.sv
Randomise method is an inbuilt virtual method which can't be modified. It is made of another two inbuilt methods called pre_randomize() and post_randomize() methods. Both of these methods are void functions. It can be overridden. When randomise is called both the methods are automatically called.
Randomize method is used to randomize the transaction.